# docs/swagger.yaml
basePath: /api/
definitions:
  handler.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.AdminCreateAccountInput:
    properties:
      balance:
        minimum: 0
        type: number
      isAdmin:
        type: boolean
      password:
        type: string
      username:
        type: string
    required:
    - balance
    - isAdmin
    - password
    - username
    type: object
  models.AdminCreateAccountOutput:
    properties:
      account:
        $ref: '#/definitions/models.GetAccountOutput'
    type: object
...
  models.AdminGetAccountOutput:
    properties:
      account:
        $ref: '#/definitions/models.GetAccountOutput'
      isDeleted:
        type: boolean
    type: object
  models.AdminGetAccountsOutput:
    properties:
      accounts:
        items:
          $ref: '#/definitions/models.GetAccountOutput'
        type: array
    type: object
...
  models.AdminUpdateAccountInput:
    properties:
      balance:
        minimum: 0
        type: number
      isAdmin:
        type: boolean
      password:
        type: string
      username:
        type: string
    required:
    - balance
    - isAdmin
    - password
    - username
    type: object
  models.AdminUpdateAccountOutput:
    properties:
      account:
        $ref: '#/definitions/models.GetAccountOutput'
    type: object
...
  models.GetAccountOutput:
    properties:
      balance:
        type: number
      id:
        type: integer
      isAdmin:
        type: boolean
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: API for transportation rental service
  title: SimbirGoAPI
  version: 1.0.0
paths:
...
  /Admin/Account:
    get:
      consumes:
      - application/json
      description: Returns a list of accounts by sampling condition.
      parameters:
      - description: '-'
        in: query
        name: count
        required: true
        type: number
      - description: '-'
        in: query
        name: start
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminGetAccountsOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Accounts information
      tags:
      - admin-account
  /Admin/Account/:
    post:
      consumes:
      - application/json
      description: Creating a new account by the administrator.
      parameters:
      - description: '-'
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AdminCreateAccountInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminCreateAccountOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Create account
      tags:
      - admin-account
  /Admin/Account/{id}:
    delete:
      consumes:
      - application/json
      description: Deleting account by id.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Delete account
      tags:
      - admin-account
    get:
      consumes:
      - application/json
      description: Returns account information.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminGetAccountOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Account information
      tags:
      - admin-account
    put:
      consumes:
      - application/json
      description: Account administrator changes account by username.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: string
      - description: '-'
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AdminUpdateAccountInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminUpdateAccountOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Update account
      tags:
      - admin-account
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"